Reverted back changes up to 1.0rc0. One more try.

+++ trunk/www/clients_view.php
@@ -2,55 +2,34 @@
 	require_once('src/prepend.inc.php');
 
 	if ($_SESSION["uid"] != 0)
-	   UI::Redirect("index.php");
+	   CoreUtils::Redirect("index.php");
 	
 	// Post actions
 	if ($_POST && $post_actionsubmit)
 	{
-		switch($post_action)
+		if ($post_action == "delete")
 		{
-			case "activate":
-			case "deactivate":
+			// Delete users
+			$i = 0;			
+			foreach ((array)$post_delete as $k=>$v)
+			{
+				$i++;
+				$db->Execute("DELETE FROM clients WHERE id='{$v}'");
 				
-				$flag = ($post_action == "activate") ? '1' : '0';
+				$farms = $db->GetAll("SELECT * FROM farms WHERE clientid='{$v}'");
+			    foreach ($farms as $farm)
+			    {
+				    $db->Execute("DELETE FROM farms WHERE id='{$farm["id"]}'");
+				    $db->Execute("DELETE FROM farm_amis WHERE farmid='{$farm["id"]}'");
+				    $db->Execute("DELETE FROM farm_instances WHERE farmid='{$farm["id"]}'");
+			    }
 				
-				$i = 0;			
-				foreach ((array)$post_delete as $clientid)
-				{
-					$db->Execute("UPDATE clients SET isactive=? WHERE id=?", array($flag, $clientid));
-					$i++;
-				}
-				
-				$mess = "{$i} clients updated";
-				UI::Redirect("clients_view.php");
-				
-				break;
+				@unlink(APPPATH . "/etc/clients_keys/{$_SESSION['uid']}/pk.pem");
+				@unlink(APPPATH . "/etc/clients_keys/{$_SESSION['uid']}/cert.pem");
+			}
 			
-			case "delete":
-				
-				// Delete users
-				$i = 0;			
-				foreach ((array)$post_delete as $clientid)
-				{
-					$i++;
-					$db->Execute("DELETE FROM clients WHERE id='{$clientid}'");
-					
-					$farms = $db->GetAll("SELECT * FROM farms WHERE clientid='{$clientid}'");
-				    foreach ($farms as $farm)
-				    {
-					    $db->Execute("DELETE FROM farms WHERE id='{$farm["id"]}'");
-					    $db->Execute("DELETE FROM farm_amis WHERE farmid='{$farm["id"]}'");
-					    $db->Execute("DELETE FROM farm_instances WHERE farmid='{$farm["id"]}'");
-				    }
-					
-					@unlink(APPPATH . "/etc/clients_keys/{$clientid}/pk.pem");
-					@unlink(APPPATH . "/etc/clients_keys/{$clientid}/cert.pem");
-				}
-				
-				$mess = "{$i} clients deleted";
-				UI::Redirect("clients_view.php");
-				
-				break;
+			$mess = "{$i} clients deleted";
+			CoreUtils::Redirect("clients_view.php");
 		}
 	}
 
@@ -91,8 +70,6 @@
 	
 	$display["page_data_options_add"] = true;
 	$display["page_data_options"] = array(
-		array("name" => "Activate", "action" => "activate"),
-		array("name" => "Deactivate", "action" => "deactivate"),
 		array("name" => "Delete", "action" => "delete")
 	);
 	require_once ("src/append.inc.php");
